STITCH-1986 - Add watch to remote mongo collection (#98)

+++ core/services/mongodb-remote/src/main/java/com/mongodb/stitch/core/services/mongodb/remote/sync/internal/DataSynchronizer.java
@@ -16,11 +16,6 @@
 
 package com.mongodb.stitch.core.services.mongodb.remote.sync.internal;
 
-import static com.mongodb.stitch.core.services.mongodb.remote.sync.internal.ChangeEvent.changeEventForLocalDelete;
-import static com.mongodb.stitch.core.services.mongodb.remote.sync.internal.ChangeEvent.changeEventForLocalInsert;
-import static com.mongodb.stitch.core.services.mongodb.remote.sync.internal.ChangeEvent.changeEventForLocalReplace;
-import static com.mongodb.stitch.core.services.mongodb.remote.sync.internal.ChangeEvent.changeEventForLocalUpdate;
-
 import com.mongodb.Block;
 import com.mongodb.Function;
 import com.mongodb.MongoClientSettings;
@@ -47,13 +42,16 @@ import com.mongodb.stitch.core.internal.common.Callback;
 import com.mongodb.stitch.core.internal.common.Dispatcher;
 import com.mongodb.stitch.core.internal.net.NetworkMonitor;
 import com.mongodb.stitch.core.services.internal.CoreStitchServiceClient;
+import com.mongodb.stitch.core.services.mongodb.remote.ChangeEvent;
+import com.mongodb.stitch.core.services.mongodb.remote.ExceptionListener;
+import com.mongodb.stitch.core.services.mongodb.remote.OperationType;
 import com.mongodb.stitch.core.services.mongodb.remote.RemoteDeleteResult;
 import com.mongodb.stitch.core.services.mongodb.remote.RemoteUpdateResult;
+import com.mongodb.stitch.core.services.mongodb.remote.UpdateDescription;
 import com.mongodb.stitch.core.services.mongodb.remote.internal.CoreRemoteMongoClient;
 import com.mongodb.stitch.core.services.mongodb.remote.internal.CoreRemoteMongoCollection;
 import com.mongodb.stitch.core.services.mongodb.remote.sync.ChangeEventListener;
 import com.mongodb.stitch.core.services.mongodb.remote.sync.ConflictHandler;
-import com.mongodb.stitch.core.services.mongodb.remote.sync.ErrorListener;
 
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
@@ -92,7 +90,7 @@ import org.bson.diagnostics.Loggers;
  */
 public class DataSynchronizer implements NetworkMonitor.StateListener {
 
-  static final String DOCUMENT_VERSION_FIELD = "__stitch_sync_version";
+  public static final String DOCUMENT_VERSION_FIELD = "__stitch_sync_version";
 
   private final CoreStitchServiceClient service;
   private final CoreRemoteMongoClient remoteClient;
@@ -118,7 +116,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
   private final Lock listenersLock;
   private final Dispatcher eventDispatcher;
 
-  private ErrorListener errorListener;
+  private ExceptionListener exceptionListener;
   private Thread initThread;
   private DispatchGroup ongoingOperationsGroup;
 
@@ -348,7 +346,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
   public <T> void configure(@Nonnull final MongoNamespace namespace,
                             @Nullable final ConflictHandler<T> conflictHandler,
                             @Nullable final ChangeEventListener<T> changeEventListener,
-                            @Nullable final ErrorListener errorListener,
+                            @Nullable final ExceptionListener exceptionListener,
                             @Nonnull final Codec<T> codec) {
     this.waitUntilInitialized();
 
@@ -360,7 +358,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
       return;
     }
 
-    this.errorListener = errorListener;
+    this.exceptionListener = exceptionListener;
 
     this.syncConfig.getNamespaceConfig(namespace).configure(
         conflictHandler,
@@ -593,7 +591,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
             syncRemoteChangeEventToLocal(
                 nsConfig,
                 docConfig,
-                changeEventForLocalReplace(
+                ChangeEvents.changeEventForLocalReplace(
                     nsConfig.getNamespace(),
                     docId,
                     latestDocumentMap.get(docId),
@@ -623,7 +621,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
           syncRemoteChangeEventToLocal(
               nsConfig,
               docConfig,
-              changeEventForLocalDelete(
+              ChangeEvents.changeEventForLocalDelete(
                   nsConfig.getNamespace(),
                   unseenId,
                   docConfig.hasUncommittedWrites()
@@ -855,7 +853,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
       resolveConflict(
               nsConfig.getNamespace(),
               docConfig,
-              changeEventForLocalDelete(
+              ChangeEvents.changeEventForLocalDelete(
                       nsConfig.getNamespace(),
                       docConfig.getDocumentId(),
                       docConfig.hasUncommittedWrites()));
@@ -914,7 +912,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
     resolveConflict(
             nsConfig.getNamespace(),
             docConfig,
-            changeEventForLocalReplace(
+            ChangeEvents.changeEventForLocalReplace(
                     nsConfig.getNamespace(),
                     docConfig.getDocumentId(),
                     newestRemoteDocument,
@@ -1137,7 +1135,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
                   continue;
                 }
 
-                final ChangeEvent.UpdateDescription localUpdateDescription =
+                final UpdateDescription localUpdateDescription =
                         localChangeEvent.getUpdateDescription();
                 if (localUpdateDescription.getRemovedFields().isEmpty()
                         && localUpdateDescription.getUpdatedFields().isEmpty()) {
@@ -1353,7 +1351,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
   private void emitError(final CoreDocumentSynchronizationConfig docConfig,
                          final String msg,
                          final Exception ex) {
-    if (this.errorListener != null) {
+    if (this.exceptionListener != null) {
       final Exception dispatchException;
       if (ex == null) {
         dispatchException = new DataSynchronizerException(msg);
@@ -1361,7 +1359,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
         dispatchException = ex;
       }
       this.eventDispatcher.dispatch(() -> {
-        errorListener.onError(docConfig.getDocumentId(), dispatchException);
+        exceptionListener.onError(docConfig.getDocumentId(), dispatchException);
         return null;
       });
     }
@@ -1414,11 +1412,10 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
     final Object resolvedDocument;
     final ChangeEvent transformedRemoteEvent;
     try {
-      final ChangeEvent transformedLocalEvent = ChangeEvent.transformChangeEventForUser(
+      final ChangeEvent transformedLocalEvent = ChangeEvents.transformChangeEventForUser(
           docConfig.getLastUncommittedChangeEvent(),
           syncConfig.getNamespaceConfig(namespace).getDocumentCodec());
-      transformedRemoteEvent =
-          ChangeEvent.transformChangeEventForUser(
+      transformedRemoteEvent = ChangeEvents.transformChangeEventForUser(
               remoteEvent,
               syncConfig.getNamespaceConfig(namespace).getDocumentCodec());
       resolvedDocument = resolveConflictWithResolver(
@@ -1447,7 +1444,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
     }
 
     final BsonDocument remoteVersion;
-    if (remoteEvent.getOperationType() == ChangeEvent.OperationType.DELETE) {
+    if (remoteEvent.getOperationType() == OperationType.DELETE) {
       // We expect there will be no version on the document. Note: it's very possible
       // that the document could be reinserted at this point with no version field and we
       // would end up deleting it, unless we receive a notification in time.
@@ -1593,9 +1590,9 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
     // a. When the document is looked up, if it cannot be found the synthesized change event is a
     // DELETE, otherwise it's a REPLACE.
     if (document == null) {
-      return changeEventForLocalDelete(ns, documentId, false);
+      return ChangeEvents.changeEventForLocalDelete(ns, documentId, false);
     }
-    return changeEventForLocalReplace(ns, documentId, document, false);
+    return ChangeEvents.changeEventForLocalReplace(ns, documentId, document, false);
   }
 
   /**
@@ -1705,14 +1702,17 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
       final BsonValue... documentIds
   ) {
     this.waitUntilInitialized();
+    boolean added = false;
 
     try {
       ongoingOperationsGroup.enter();
       for (final BsonValue documentId : documentIds) {
-        syncConfig.addSynchronizedDocument(namespace, documentId);
+        added = syncConfig.addSynchronizedDocument(namespace, documentId) || added;
       }
 
-      triggerListeningToNamespace(namespace);
+      if (added) {
+        triggerListeningToNamespace(namespace);
+      }
     } finally {
       ongoingOperationsGroup.exit();
     }
@@ -1732,20 +1732,26 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
     this.waitUntilInitialized();
     final Lock lock = this.syncConfig.getNamespaceConfig(namespace).getLock().writeLock();
     lock.lock();
+    boolean removed = false;
     try {
       ongoingOperationsGroup.enter();
+
       for (final BsonValue documentId : documentIds) {
-        syncConfig.removeSynchronizedDocument(namespace, documentId);
+        removed = syncConfig.removeSynchronizedDocument(namespace, documentId) || removed;
       }
 
-      getLocalCollection(namespace).deleteMany(
-          new Document("_id", new Document("$in", Arrays.asList(documentIds))));
+      if (removed) {
+        getLocalCollection(namespace).deleteMany(
+            new Document("_id", new Document("$in", Arrays.asList(documentIds))));
+      }
     } finally {
       lock.unlock();
       ongoingOperationsGroup.exit();
     }
 
-    triggerListeningToNamespace(namespace);
+    if (removed) {
+      triggerListeningToNamespace(namespace);
+    }
   }
 
   /**
@@ -1932,8 +1938,8 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
       try {
         getLocalCollection(namespace).insertOne(docForStorage);
         documentId = BsonUtils.getDocumentId(docForStorage);
-        event = changeEventForLocalInsert(namespace, docForStorage, true);
-        final CoreDocumentSynchronizationConfig config = syncConfig.addSynchronizedDocument(
+        event = ChangeEvents.changeEventForLocalInsert(namespace, docForStorage, true);
+        final CoreDocumentSynchronizationConfig config = syncConfig.addAndGetSynchronizedDocument(
             namespace,
             documentId
         );
@@ -1975,8 +1981,8 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
         for (final BsonDocument document : docsForStorage) {
           final BsonValue documentId = BsonUtils.getDocumentId(document);
           final ChangeEvent<BsonDocument> event =
-              changeEventForLocalInsert(namespace, document, true);
-          final CoreDocumentSynchronizationConfig config = syncConfig.addSynchronizedDocument(
+              ChangeEvents.changeEventForLocalInsert(namespace, document, true);
+          final CoreDocumentSynchronizationConfig config = syncConfig.addAndGetSynchronizedDocument(
               namespace,
               documentId
           );
@@ -2083,15 +2089,15 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
         // else this is an update
         if (documentBeforeUpdate == null && updateOptions.isUpsert()) {
           triggerNamespace = true;
-          config = syncConfig.addSynchronizedDocument(namespace, documentId);
-          event = changeEventForLocalInsert(namespace, documentAfterUpdate, true);
+          config = syncConfig.addAndGetSynchronizedDocument(namespace, documentId);
+          event = ChangeEvents.changeEventForLocalInsert(namespace, documentAfterUpdate, true);
         } else {
           triggerNamespace = false;
           config = syncConfig.getSynchronizedDocument(namespace, documentId);
-          event = changeEventForLocalUpdate(
+          event = ChangeEvents.changeEventForLocalUpdate(
               namespace,
               BsonUtils.getDocumentId(documentAfterUpdate),
-              ChangeEvent.UpdateDescription.diff(documentBeforeUpdate, documentAfterUpdate),
+              UpdateDescription.diff(documentBeforeUpdate, documentAfterUpdate),
               documentAfterUpdate,
               true);
         }
@@ -2218,14 +2224,14 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
               // treat the upsert as an insert, as far as sync is concerned
               // else treat it as a standard update
               if (beforeDocument == null && updateOptions.isUpsert()) {
-                config = syncConfig.addSynchronizedDocument(namespace, documentId);
-                event = changeEventForLocalInsert(namespace, afterDocument, true);
+                config = syncConfig.addAndGetSynchronizedDocument(namespace, documentId);
+                event = ChangeEvents.changeEventForLocalInsert(namespace, afterDocument, true);
               } else {
                 config = syncConfig.getSynchronizedDocument(namespace, documentId);
-                event = changeEventForLocalUpdate(
+                event = ChangeEvents.changeEventForLocalUpdate(
                     namespace,
                     documentId,
-                    ChangeEvent.UpdateDescription.diff(beforeDocument, afterDocument),
+                    UpdateDescription.diff(beforeDocument, afterDocument),
                     afterDocument,
                     true);
               }
@@ -2293,13 +2299,13 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
               docForStorage,
               new FindOneAndReplaceOptions().upsert(true).returnDocument(ReturnDocument.AFTER));
 
-      if (remoteEvent.getOperationType() == ChangeEvent.OperationType.DELETE) {
-        event = changeEventForLocalInsert(namespace, documentAfterUpdate, true);
+      if (remoteEvent.getOperationType() == OperationType.DELETE) {
+        event = ChangeEvents.changeEventForLocalInsert(namespace, documentAfterUpdate, true);
       } else {
-        event = changeEventForLocalUpdate(
+        event = ChangeEvents.changeEventForLocalUpdate(
             namespace,
             documentId,
-            ChangeEvent.UpdateDescription.diff(
+            UpdateDescription.diff(
                     sanitizeDocument(remoteEvent.getFullDocument()),
                     documentAfterUpdate),
             docForStorage,
@@ -2366,7 +2372,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
         undoCollection.deleteOne(getDocumentIdFilter(documentId));
       }
 
-      event = changeEventForLocalReplace(namespace, documentId, docForStorage, false);
+      event = ChangeEvents.changeEventForLocalReplace(namespace, documentId, docForStorage, false);
     } finally {
       lock.unlock();
     }
@@ -2413,12 +2419,12 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
         undoCollection.insertOne(docToDelete);
 
         result = localCollection.deleteOne(filter);
-        event = changeEventForLocalDelete(namespace, documentId, true);
+        event = ChangeEvents.changeEventForLocalDelete(namespace, documentId, true);
 
         // this block is to trigger coalescence for a delete after insert
         if (config.getLastUncommittedChangeEvent() != null
             && config.getLastUncommittedChangeEvent().getOperationType()
-            == ChangeEvent.OperationType.INSERT) {
+            == OperationType.INSERT) {
           desyncDocumentsFromRemote(config.getNamespace(), config.getDocumentId());
           undoCollection.deleteOne(getDocumentIdFilter(config.getDocumentId()));
           return result;
@@ -2480,12 +2486,12 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
           }
 
           final ChangeEvent<BsonDocument> event =
-              changeEventForLocalDelete(namespace, documentId, true);
+              ChangeEvents.changeEventForLocalDelete(namespace, documentId, true);
 
           // this block is to trigger coalescence for a delete after insert
           if (config.getLastUncommittedChangeEvent() != null
               && config.getLastUncommittedChangeEvent().getOperationType()
-              == ChangeEvent.OperationType.INSERT) {
+              == OperationType.INSERT) {
             desyncDocumentsFromRemote(config.getNamespace(), config.getDocumentId());
             undoCollection.deleteOne(getDocumentIdFilter(documentId));
             continue;
@@ -2539,7 +2545,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
       }
       localCollection
           .deleteOne(getDocumentIdFilter(documentId));
-      event = changeEventForLocalDelete(namespace, documentId, true);
+      event = ChangeEvents.changeEventForLocalDelete(namespace, documentId, true);
       config.setSomePendingWrites(
           logicalT, atVersion, event);
       if (documentToDelete != null) {
@@ -2586,7 +2592,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
     } finally {
       lock.unlock();
     }
-    emitEvent(documentId, changeEventForLocalDelete(namespace, documentId, false));
+    emitEvent(documentId, ChangeEvents.changeEventForLocalDelete(namespace, documentId, false));
   }
 
   private void triggerListeningToNamespace(final MongoNamespace namespace) {
@@ -2659,7 +2665,7 @@ public class DataSynchronizer implements NetworkMonitor.StateListener {
             if (namespaceListener.getEventListener() != null) {
               namespaceListener.getEventListener().onEvent(
                   documentId,
-                  ChangeEvent.transformChangeEventForUser(
+                  ChangeEvents.transformChangeEventForUser(
                       event, namespaceListener.getDocumentCodec()));
             }
           } catch (final Exception ex) {
